--- docs/examples/static_files/upgrade_from_static_1.py
@@ -1,8 +1,0 @@
-from litestar import Litestar
-from litestar.static_files.config import StaticFilesConfig
-
-app = Litestar(
-    static_files_config=[
-        StaticFilesConfig(directories=["assets"], path="/static"),
-    ],
-)
