--- docs/examples/static_files/upgrade_from_static_2.py
@@ -1,8 +1,0 @@
-from litestar import Litestar
-from litestar.static_files import create_static_files_router
-
-app = Litestar(
-    route_handlers=[
-        create_static_files_router(directories=["assets"], path="/static"),
-    ],
-)
