+++ docs/release-notes/whats-new-2.rst
@@ -93,7 +93,7 @@ Imports
 +----------------------------------------------------+------------------------------------------------------------------------+
 | ``starlite.OpenAPIConfig``                         | :class:`.openapi.OpenAPIConfig`                                        |
 +----------------------------------------------------+------------------------------------------------------------------------+
-| ``starlite.StaticFilesConfig``                     | :class:`.static_files.config.StaticFilesConfig`                        |
+| ``starlite.StaticFilesConfig``                     | ``.static_files.config.StaticFilesConfig``                             |
 +----------------------------------------------------+------------------------------------------------------------------------+
 | ``starlite.TemplateConfig``                        | :class:`.template.TemplateConfig`                                      |
 +----------------------------------------------------+------------------------------------------------------------------------+
