+++ litestar/_asgi/routing_trie/traversal.py
@@ -150,8 +150,7 @@ def parse_path_to_route(
             if not any(remaining_path.startswith(f"{sub_route}/") for sub_route in children):
                 asgi_app, handler = parse_node_handlers(node=mount_node, method=method)
                 remaining_path = remaining_path or "/"
-                if not mount_node.is_static:
-                    remaining_path = remaining_path if remaining_path.endswith("/") else f"{remaining_path}/"
+                remaining_path = remaining_path if remaining_path.endswith("/") else f"{remaining_path}/"
                 return asgi_app, handler, remaining_path, {}, root_node.path_template
 
         node, path_parameters, path = traverse_route_map(
