+++ litestar/cli/_utils.py
@@ -388,15 +388,6 @@ def show_app_info(app: Litestar) -> None:  # pragma: no cover
     if app.template_engine:
         table.add_row("Template engine", type(app.template_engine).__name__)
 
-    if app.static_files_config:
-        static_files_configs = app.static_files_config
-        static_files_info = [
-            f"path=[yellow]{static_files.path}[/] dirs=[yellow]{', '.join(map(str, static_files.directories))}[/] "
-            f"html_mode={_format_is_enabled(static_files.html_mode)}"
-            for static_files in static_files_configs
-        ]
-        table.add_row("Static files", "\n".join(static_files_info))
-
     middlewares = []
     for middleware in app.middleware:
         updated_middleware = middleware.middleware if isinstance(middleware, DefineMiddleware) else middleware
