--- litestar/static_files/__init__.py
@@ -1,4 +1,0 @@
-from litestar.static_files.base import StaticFiles
-from litestar.static_files.config import StaticFilesConfig, create_static_files_router
-
-__all__ = ("StaticFiles", "StaticFilesConfig", "create_static_files_router")
