+++ litestar/testing/helpers.py
@@ -27,7 +27,6 @@ if TYPE_CHECKING:
     from litestar.openapi.config import OpenAPIConfig
     from litestar.openapi.spec import SecurityRequirement
     from litestar.plugins import PluginProtocol
-    from litestar.static_files.config import StaticFilesConfig
     from litestar.stores.base import Store
     from litestar.stores.registry import StoreRegistry
     from litestar.template.config import TemplateConfig
@@ -103,7 +102,6 @@ def create_test_client(
     signature_namespace: Mapping[str, Any] | None = None,
     signature_types: Sequence[Any] | None = None,
     state: State | None = None,
-    static_files_config: Sequence[StaticFilesConfig] | None = None,
     stores: StoreRegistry | dict[str, Store] | None = None,
     tags: Sequence[str] | None = None,
     template_config: TemplateConfig | None = None,
@@ -226,7 +224,6 @@ def create_test_client(
         signature_types: A sequence of types for use in forward reference resolution during signature modeling.
             These types will be added to the signature namespace using their ``__name__`` attribute.
         state: An optional :class:`State <.datastructures.State>` for application state.
-        static_files_config: A sequence of :class:`StaticFilesConfig <.static_files.StaticFilesConfig>`
         stores: Central registry of :class:`Store <.stores.base.Store>` that will be available throughout the
             application. If this is a dictionary to it will be passed to a
             :class:`StoreRegistry <.stores.registry.StoreRegistry>`. If it is a
@@ -292,7 +289,6 @@ def create_test_client(
         signature_namespace=signature_namespace,
         signature_types=signature_types,
         state=state,
-        static_files_config=static_files_config,
         stores=stores,
         tags=tags,
         template_config=template_config,
@@ -364,7 +360,6 @@ def create_async_test_client(
     signature_namespace: Mapping[str, Any] | None = None,
     signature_types: Sequence[Any] | None = None,
     state: State | None = None,
-    static_files_config: Sequence[StaticFilesConfig] | None = None,
     stores: StoreRegistry | dict[str, Store] | None = None,
     tags: Sequence[str] | None = None,
     template_config: TemplateConfig | None = None,
@@ -487,7 +482,6 @@ def create_async_test_client(
         signature_types: A sequence of types for use in forward reference resolution during signature modeling.
             These types will be added to the signature namespace using their ``__name__`` attribute.
         state: An optional :class:`State <.datastructures.State>` for application state.
-        static_files_config: A sequence of :class:`StaticFilesConfig <.static_files.StaticFilesConfig>`
         stores: Central registry of :class:`Store <.stores.base.Store>` that will be available throughout the
             application. If this is a dictionary to it will be passed to a
             :class:`StoreRegistry <.stores.registry.StoreRegistry>`. If it is a
@@ -552,7 +546,6 @@ def create_async_test_client(
         signature_namespace=signature_namespace,
         signature_types=signature_types,
         state=state,
-        static_files_config=static_files_config,
         stores=stores,
         tags=tags,
         template_config=template_config,
