+++ test_apps/static_files_test_app/main.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from litestar import Litestar, get
-from litestar.static_files.config import StaticFilesConfig
+from litestar.static_files.config import create_static_files_router
 
 
 @get("/")
@@ -10,9 +10,8 @@ async def handler() -> dict[str, str]:
 
 
 app = Litestar(
-    route_handlers=[],
-    static_files_config=[
-        StaticFilesConfig(directories=[Path(__file__).parent / "public"], path="/", html_mode=True),
+    route_handlers=[
+        create_static_files_router(directories=[Path(__file__).parent / "public"], path="/", html_mode=True),
     ],
 )
